Refactored bulk copy worker to enhance modularity and clarity by introducing new utility functions for batch management and status updates, while removing redundant code and improving error handling.
